Generate the correct SERVER_PROTOCOL_VERSION.
Closes in development: #836911
Closes: #836566

SERVER_PROTOCOL_VERSION is used to check if the running server's
and the client's version match. It is set in
include/wine/server_protocol.h (490 in 1.8.4), and increased by
1 by make_requests. But we clean server_protocol.h because it is
generated, so SERVER_PROTOCOL_VERSION in Debian is always 1.

To fix this retrieve SERVER_PROTOCOL_VERSION on upstream import
before server_protocol.h is cleaned, and store it (decreased by
1) in request.patch. make_request will then regenerate
server_protocol.h with the correct SERVER_PROTOCOL_VERSION.

# Conflicts:
#	debian/changelog
#	debian/patches/generate/request.patch

Note cherrypick: Change initial version in requests.patch to 490
and adapt bug numbers.